Services refactor

+++ Valour/Server/Api/Dynamic/PlanetRoleApi.cs
@@ -60,6 +60,8 @@ public class PlanetRoleApi
         PlanetRoleService roleService)
     {
         var oldRole = await roleService.GetAsync(id);
+        if (oldRole is null)
+            return ValourResult.NotFound("Role not found.");
 
         // Get member
         var member = await memberService.GetCurrentAsync(oldRole.PlanetId);
@@ -68,6 +70,9 @@ public class PlanetRoleApi
 
         if (!await memberService.HasPermissionAsync(member, PlanetPermissions.ManageRoles))
             return ValourResult.LacksPermission(PlanetPermissions.ManageRoles);
+
+        if (await memberService.GetAuthorityAsync(member) <= role.GetAuthority())
+            return ValourResult.Forbid("You can only edit roles under your own.");
 
         var result = await roleService.PutAsync(role);
         if (!result.Success)
@@ -84,6 +89,8 @@ public class PlanetRoleApi
         PlanetRoleService roleService)
     {
         var role = await roleService.GetAsync(id);
+        if (role is null)
+            return ValourResult.NotFound("Role not found.");
 
         // Get member
         var member = await memberService.GetCurrentAsync(role.PlanetId);
@@ -93,11 +100,40 @@ public class PlanetRoleApi
         if (!await memberService.HasPermissionAsync(member, PlanetPermissions.ManageRoles))
             return ValourResult.LacksPermission(PlanetPermissions.ManageRoles);
 
+        if (await memberService.GetAuthorityAsync(member) <= role.GetAuthority())
+            return ValourResult.Forbid("You can only delete roles under your own.");
+
         var result = await roleService.DeleteAsync(role);
         if (!result.Success)
             return ValourResult.Problem(result.Message);
 
         return Results.NoContent();
 
+    }
+
+    /// <summary>
+    /// This returns the base permission nodes for this role, if they exist
+    /// </summary>
+    [ValourRoute(HttpVerbs.Get, "api/roles/{id}/permissions")]
+    [UserRequired(UserPermissionsEnum.Membership)]
+    public static async Task<IResult> GetPermissionsAsync(
+        long id,
+        PlanetRoleService roleService,
+        PlanetMemberService memberService)
+    {
+        // Get the role
+        var role = await roleService.GetAsync(id);
+        if (role is null)
+            return ValourResult.NotFound("Role not found");
+
+        // Get member
+        var member = await memberService.GetCurrentAsync(role.PlanetId);
+        if (member is null)
+            return ValourResult.NotPlanetMember();
+
+        var nodes = await roleService.GetBasePermissionNodesAsync(id);
+
+        // Return json
+        return Results.Json(nodes);
     }
 }
